fix(auth): move logout to useAuth and fix notification toast placement
