+++ src/app/login/page.tsx
@@ -111,7 +111,7 @@ export default function LoginPage() {
         </CardContent>
       </Card>
 
-      <NotificationToast />
+      
     </div>
   );
 }
