+++ src/components/common/user-profile-button.tsx
@@ -3,12 +3,14 @@
 import { useState, useRef, useEffect } from "react"
 import { Button } from "@/components/ui/button"
 import { Badge } from "@/components/ui/badge"
+import { useAuth } from "@/features/auth/hooks/useAuth"
 // import { Settings, HelpCircle, LogOut, ChevronDown, UserCircle, Palette } from "lucide-react"
-import { useAuth } from "@/contexts/auth-context"
+interface UserProfileButtonProps {
+  onLogout: () => void
+  onSettings?: () => void
+}
 
-
-
-export function UserProfileButton() {
+export function UserProfileButton({ onLogout, onSettings }: UserProfileButtonProps) {
   const [isOpen, setIsOpen] = useState(false)
   const { state: authState } = useAuth()
   const dropdownRef = useRef<HTMLDivElement>(null)
